Move default tests to new TcpOptionsDefaultTests suite
